project model and handle project test data

+++ models/user.go
@@ -2,7 +2,7 @@ package models
 
 // User developer register and signin
 type User struct {
-	DevID       int    `json:"dev_id"`
+	DevID       int64  `json:"dev_id"`
 	DevEmail    string `json:"dev_email"`
 	DevPassword string `json:"dev_password"`
 }
